fix(cli): hotswap deploy fails on multiple CfnEvaluationException (#22339)
closes #22323 

To avoid unhandled rejections, we run promises just before we call `Promise.all`.

The concern of this fix is that hotswap process may take longer time because now async tasks run lazily. However I don't think it will be a big problem since those tasks are not I/O bound, so most of them are already running sequentially, not in parallel.

----

### All Submissions:

* [X] Have you followed the guidelines in our [Contributing guide?](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md)

### Adding new Unconventional Dependencies:

* [ ] This PR adds new unconventional dependencies following the process described [here](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md/#adding-new-unconventional-dependencies)

### New Features

* [ ] Have you added the new feature to an [integration test](https://github.com/aws/aws-cdk/blob/main/INTEGRATION_TESTS.md)?
	* [ ] Did you use `yarn integ` to deploy the infrastructure and generate the snapshot (i.e. `yarn integ` without `--dry-run`)?

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*