chore(lambda-python): remove arm integ tests
It is not possible to run the ARM integ tests on x86 machines (i.e.
CodeBuild). This is currently causing our jsii-main pipeline to fail.

I believe this is not causing our aws-cdk pipeline to fail due to image
caching. I tested this locally and did not receive the failure until I
ran `docker system prune`.

The error that you receive is:

```
The requested image's platform (linux/arm64/v8) does not match the detected host platform (linux/amd64) and no specific platform was requested
```