chore(aws-ecs): add service connect changes to port mapping and task def (#82)
This commit adds the fields `name` and `appProtocol` to Port Mapping, along with validation and unit tests to ensure that the new fields don't conflict in confusing ways for customers. 

This also adds validation at the task definition level to ensure no two port mappings on different containers have the same name.