chore(aws-eks): pin `aws-cdk-lib` devDependency to `2.47.0`
The v2 release build is currently failing due to a cycle between
`aws-cdk-lib -> @aws-cdk/aws-eks -> aws-cdk-lib`. This only occurs
during the release build (and not the PR build) because we run
`align-versions.sh` in the release build. I tested this locally and the
error went away after pinning the version.