fix(eks): kubectl get handler output includes stderr (#22658)
The custom resource that handles kubectl requests returns standard output and standard error in the same string. This is normally okay, because we likely get _either_ a successful result with standard output, or an unsuccessful one with standard error. However, there are certain circumstances where kubectl returns a successful result with standard output _as well as_ warnings in standard error. These warnings are harmless, but because of how the custom resource is set up, it corrupts the output that gets returned. This results in integ test failure. 

These changes separate out standard output and standard error that gets returned by kubectl, and only prints the error if we mean to throw an error.

----

### All Submissions:

* [ ] Have you followed the guidelines in our [Contributing guide?](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md)

### Adding new Unconventional Dependencies:

* [ ] This PR adds new unconventional dependencies following the process described [here](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md/#adding-new-unconventional-dependencies)

### New Features

* [ ] Have you added the new feature to an [integration test](https://github.com/aws/aws-cdk/blob/main/INTEGRATION_TESTS.md)?
	* [ ] Did you use `yarn integ` to deploy the infrastructure and generate the snapshot (i.e. `yarn integ` without `--dry-run`)?

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*