chore(core): inject `TreeMetadata` resource at the end (during synthesis)
The `TreeMetadata` resource was being created as part of the `App`
constructor, which meant that you could not add any context to the App
node (you must add context _before_ any constructs are added). This
moves the creation of `TreeMetadata` to be injected during `synthesis`
similar to how we currently inject the metadata resource.

A side effect of this change is that we will now include tree metadata
for resources that are added in `Stack.synthesize()` methods.
The [synthesizeTree](https://github.com/aws/aws-cdk/blob/95e9e0f54a1dfb628d5146bd9b69c292547cf090/packages/@aws-cdk/core/lib/private/synthesis.ts#L193-L210)
function visits nodes in the order that they were added to the tree.
Previously the `TreeMetadata` was the first node so the TreeMetadata was
generated first. Then if any Stacks would call `Stack.synthesize()` and
add additional constructs (e.g. `BootstrapVersion`) these would not be
added to the tree.