fix(cli): assetParallelism option in cdk.json is not recognized
Version 2.48.0 introduced a flag to enable/disable parallelism when
building and publishing assets (see #22579 ). The flag is supposed to
be set via either environment variable (`CDK_ASSET_PARALLELISM`), command
line argument (`-[no]-asset-parallelism`), or the `cdk.json` file. However,
the last method does not work. We are not passing the option from that
file properly.