chore(iam): add a `PrecreatedRole` class
This PR adds a new private class for creating a `PrecreatedRole`. For
background see the [precreated-roles
RFC](https://github.com/aws/aws-cdk-rfcs/blob/master/text/0063-precreated-roles.md).

This PR adds the class and the functionality to generate a report as
part of synthesis, but it does not yet expose this publicly. A future PR
will use this class as part of the `Role` class.

The report that is generated will contain references to resources that
have not yet been created. Those values are not easy to read in a report
and make it harder to search/replace with a hard coded value once the
resource _is_ created. To account for this we replace any references
with easier to read values. For example, this:

```
"Resource": {
  "Fn::Join": [
    "",
    [
      "arn:",
      {
        "Ref": "AWS::Partition"
      },
      ":iam::",
      {
        "Ref": "AWS::AccountId"
      },
      ":role/Role"
    ]
  ]
}
```

Will become:

`"Resource": "arn:(PARTITION):iam::(ACCOUNT):role/Role"`

note: labeled as `chore` since this does not add any public facing API.
Future PRs will add the public API that will expose the report
generation

closes #22750