chore(iam): add a `PrecreatedRole` class (#22824)
This PR adds a new private class for creating a `PrecreatedRole`. For background see the [precreated-roles
RFC](https://github.com/aws/aws-cdk-rfcs/blob/master/text/0063-precreated-roles.md).

This PR adds the class and the functionality to generate a report as part of synthesis, but it does not yet expose this publicly. A future PR will use this class as part of the `Role` class.

The report that is generated will contain references to resources that have not yet been created. Those values are not easy to read in a report and make it harder to search/replace with a hard coded value once the resource _is_ created. To account for this we replace any references with easier to read values. For example, this:

```
"Resource": {
  "Fn::Join": [
    "",
    [
      "arn:",
      {
        "Ref": "AWS::Partition"
      },
      ":iam::",
      {
        "Ref": "AWS::AccountId"
      },
      ":role/Role"
    ]
  ]
}
```

Will become:

`"Resource": "arn:(PARTITION):iam::(ACCOUNT):role/Role"`

note: labeled as `chore` since this does not add any public facing API. Future PRs will add the public API that will expose the report generation

closes #22750


----

### All Submissions:

* [ ] Have you followed the guidelines in our [Contributing guide?](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md)

### Adding new Unconventional Dependencies:

* [ ] This PR adds new unconventional dependencies following the process described [here](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md/#adding-new-unconventional-dependencies)

### New Features

* [ ] Have you added the new feature to an [integration test](https://github.com/aws/aws-cdk/blob/main/INTEGRATION_TESTS.md)?
	* [ ] Did you use `yarn integ` to deploy the infrastructure and generate the snapshot (i.e. `yarn integ` without `--dry-run`)?

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*