chore(cx-api): automatically determine new feature flag version
With the new feature flags mechanism, you have to put the `v2` version
a flag becomes available in when adding a flag.

One problem: the code hasn't been released yet, so you don't know what
version that will be. You can guess it's going to be the next release,
but what if the review takes a long time, or a new release was juuuust
started by the oncall?

----------------------

Solution: we force you to put in the magic version marker `'V2NEXT'`.

We update the sorting mechanism to sort this always after all v2
versions and before any v3 versions, so it appears at the right
place in the report.

At release time, when the actual version number is determined and files
are updated and committed back to the repository, we replace the magic
marker with the actual version number and regenerate the flag report,
so that the actual release version number is put in.

These commits will find their way to the `main` branch when the
mergeback commit is created at the end of the release pipeline.