fix(iam): service principals use unnecessary exceptions (under feature flag) (#22819)
We have a database of service principal exceptions in different regions.

This database is no longer necessary: all services now use the global service principal name for in-region references, and sometimes a standardized regional service principal name for cross-opt-in-region references.

This PR changes the following things:

```ts
new ServicePrincipal('service.amazonaws.com')
// ➡️  always resolves to 'service.amazonaws.com', regardless of region
//    or service principal

new ServicePrincipal('service.amazonaws.com', { region: 'me-south-1' })
// ➡️ resolves to 'service.me-south-1.amazonaws.com' in case of a
//   cross-region reference, or just 'service.amazonaws.com' otherwise.
```

Because change is scary (and because we are only 99% sure that this change has made it to all ADC regions), we put the new behavior behind a feature flag:

```json
{
  "context": {
    "@aws-cdk/aws-iam:standardizedServicePrincipals": true
  }
}
```


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*