fix: CDK does not work in FIPS-restricted environments
In an environment where node is compiled with FIPS restrictions in mind,
`crypto.createHash('md5')` does not work, for fear of MD5 being used for
crypto purposes.

We do not use it for crypto purposes, just to come up with unique
identifiers for certain constructs. Nevertheless, CDK cannot work if
`md5` is not available from the Node standard library.

Fall back to a pure JavaScript implementation if the built-in MD5 hash
does not work. This results in about a ~10x performance penalty, but the
only thing we can do in order to produce the same templates.