feat(iam): apply permissions boundary to a `Stage`
This PR should be not be merged until #22792 which adds support for
creating a default permissions boundary as part of CDK bootstrap.

Once that is merged we will be creating a default permissions boundary
managed policy with a standard naming convention. That allows us to
easily attach this default permissions boundary to applications at
either the `Stage` or `Stack` level.

For the implementation, there are two components to "applying" the
permissions boundary.

The first is "applying" the permissions boundary to a construct scope.
For this I've used context since context has built in resolution up the
tree. For example, you could theoretically do something like the below.

```ts
// apply at the app scope
const app = new App({
  context: {
    [PERMISSIONS_BOUNDARY_CONTEXT_KEY]: {
      name: 'app-level-pb',
    },
  },
});

// apply different value at this stage
const prodStage = new Stage(app, 'Stage', {
  permissionsBoundary: PermissionsBoundary.default(),
});

const stack = new Stack(prodStage, 'Stack', {
  permissionsBoundary: PermissionsBoundary.fromName('stack-level-pb'),
});

class MyConstruct extends Construct {
  constructor(scope: Construct, id: string) {
    super(scope, id);
    const pb = PermissionsBoundary.fromName('construct-level-pb');
    pb.bind(this);
  }
}
```

While the above is possible, the most likely scenario is that you would
apply a permission boundary at the `Stage` or `Stack` level. Stages/Stacks correspond to
AWS environments (account + region) and so they would most likely also
map to bootstrap environments. Because of this I've added parameters a
new `permissionsBoundary` property to both the `Stage` & `Stack` props
to make it easier to configure.

The second aspect to "applying" the boundary is actually attaching the
permissions boundary to each IAM Role & User that is created. For this I
add an Aspect by default to every `Stack`. This aspect will look for the
`context` that was applied earlier for whether or not it should attach a
permissions boundary to the Role or User.

closes #22745