docs(aws-cloudwatch): Fix example for `MathExpression` (#22933)
In the commit f37991c4de21ecbc7db5117de7f9adc1ecc40f2b the example was changed to use a different metric to make it compile. Unfortunately the `id` for the metric was only updated in its use not in the definition.

This can be a bit confusing when reading the example for the first time and if used as is it would not result in the intended behavior. Therefore I updated the example.

----

### All Submissions:

* [x] Have you followed the guidelines in our [Contributing guide?](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md)

### Adding new Unconventional Dependencies:

* [ ] This PR adds new unconventional dependencies following the process described [here](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md/#adding-new-unconventional-dependencies)

### New Features

* [ ] Have you added the new feature to an [integration test](https://github.com/aws/aws-cdk/blob/main/INTEGRATION_TESTS.md)?
	* [ ] Did you use `yarn integ` to deploy the infrastructure and generate the snapshot (i.e. `yarn integ` without `--dry-run`)?

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*