fix(core): the string 'undefined' is recognized as a valid partition value
When the partition value is the literal string 'undefined' (because it
was stringified from the context), it should be treated as if it is
actually undefined. When looking up a fact value, the two values are already
considered the same in one code path, but are then treated differently
in the immediate logic. This is causing a new integration test to fail.

This commit makes the logic consistent.