fix(core): the string 'undefined' is recognized as a valid partition when looking up for fact values (#23023)
When the partition value is the literal string 'undefined' (because it was stringified from the context), it should be treated as if it is actually `undefined`. When looking up a fact value, the two values are already considered the same in one code path, but are then treated differently in the immediate logic. This is causing a new integration test to fail.

This commit makes the logic consistent.


----

### All Submissions:

* [X] Have you followed the guidelines in our [Contributing guide?](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md)

### Adding new Unconventional Dependencies:

* [ ] This PR adds new unconventional dependencies following the process described [here](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md/#adding-new-unconventional-dependencies)

### New Features

* [ ] Have you added the new feature to an [integration test](https://github.com/aws/aws-cdk/blob/main/INTEGRATION_TESTS.md)?
	* [ ] Did you use `yarn integ` to deploy the infrastructure and generate the snapshot (i.e. `yarn integ` without `--dry-run`)?

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*