fix(route53): cross-account delegation broken in opt-in regions
This is a revert of #22370. The fix proposed there solved one very
specific (but rare) use case, in exchange for breaking common use cases.

- *It fixed the case of*: AWS service authors using their own global
  service principal in the delegation role, with both source and target
  account opted into the region.
- *It broke the case of*: all teams that didn't have both accounts opted
  into the region.

The second case is much more common, so revert to the old behavior.

Since the regional behavior might still be useful to *some* people
somewhere, it has been relegated to a context key,
`@aws-cdk/aws-route53:useRegionalStsEndpoint`, instead. It can be
configured, but is not advertised as 99.9% of users will not need
this behavior.

Since both STS and Route53 are global and regular customers cannot
usefully use Service Principals in this particular trust policy anyway,
there is no impact to regular customers.

Fixes #23081.