feat(cloudwatch): `Stats` factory class for metric strings
The `Statistic` enum type was incorrectly publicly exposed (it should
only have been visible internally to the package), and was enhanced in
PR #23074 to have more enum values such as `P10`, `P50`, `P99_9`, etc.

The stringification of this `Statistic` type would only have worked in
TypeScript anyway (in JSII languages like Java and Python we cannot rely
on the string values of enums), and the fact that enums cannot be
parameterized made it so that we used to have a lot of redundant enum
values.

Introduce a new factory class, `Stats`, whose sole purpose is to
produce formatted strings to use as CloudWatch `statistic` values,
and advertise the use of this class.

(We probably shouldn't have been using `string` as the type in the first
place, but given that we are factories to produce them seems to be the
next best thing).