fix(toolkit): endless wait if CDKToolkit stack is `REVIEW_IN_PROGRESS`
If a `cdk bootstrap` operation is interrupted after the ChangeSet has
been created, but before it is executed, the CDKToolkit stack will be
left in the `REVIEW_IN_PROGRESS` state, which is a stable state until
something or someone either executes or deletes the pending ChangeSet.

This resulted in an endless (and silent, unless `cdk bootstrap` is
executed in verbose mode) wait, as this was considered an "in progress"
(aka unstable) state. This changes `REVIEW_IN_PROGRESS` to be treated as
stable, while still debug-logging an indication that we are "ignoring"
the existing ChangeSet.