feat(lambda-go): allow configuration of GOPROXY (#23257)
> This is a back port of #23171 to AWS CDK v1.x, motivated by impact of the same issue on v1.

We used to require direct Go package access, because Go proxies may be blocked by corporate network policies (and wouldn't you know it, it actually is at our particular workplace 🙃).

This produces a good bit of instability in our CI builds though, as `gopkg.in` website which is used to reference some of our transitive dependencies is regularly experiencing downtime.

Make Go proxies configurable and switch them back on in CI builds.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*