chore: save exclusion rules when bumping the CFN spec (#22788)
`bump-cfnspec.sh` now does a full build, with the env variable `AWSLINT_SAVE=true`. This has the effect of adding awslint exclusion rules to any package that violates awslint rules after it is created (e.g., new attributes in CFN that don't yet exist in the corresponding L2). As part of the script, these changes will be commited to Git.

Tested locally on a clean build (except for a change in the `package.json` of a construct module to check if the changes were reverted after the build) and on the CodeBuild job, by pointing it to this branch.

----

### All Submissions:

* [X] Have you followed the guidelines in our [Contributing guide?](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md)

### Adding new Unconventional Dependencies:

* [ ] This PR adds new unconventional dependencies following the process described [here](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md/#adding-new-unconventional-dependencies)

### New Features

* [ ] Have you added the new feature to an [integration test](https://github.com/aws/aws-cdk/blob/main/INTEGRATION_TESTS.md)?
	* [ ] Did you use `yarn integ` to deploy the infrastructure and generate the snapshot (i.e. `yarn integ` without `--dry-run`)?

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*